requirements.txt
# Data Collection
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2

# Data Processing
pandas>=2.0.0
numpy>=1.21.0
yfinance>=0.2.0

# NLP and Sentiment Analysis
nltk>=3.8
textblob>=0.17.1
vaderSentiment>=3.3.2
transformers>=4.21.0
torch>=1.13.0

# Visualization and Dashboard
plotly>=5.0.0
dash>=2.10.0
dash-bootstrap-components>=1.4.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
